feat(ts): borsh -> ts mvp (#1183)
for #1182
Added some basic borsh to ts generator, some notes:
- see **nord.ts** for example
- need migrate nord too after from proto, not only do to new gate.
because 1 bad serde is better than 1 bad and 1 good
- does not handle external crates conflicts(so never happened) nor doc
comments (need own proc syn macro for that or external rust parser), TS
code has no comments for now.
https://github.com/near/borsh-rs/issues/324
- code is not optimized, a lot of todo and refactoring, so improvements
later, bugs/testing. but MVP solution is here to start using.

If we merge it, we are bullish on borsh, so can patch as needed for
action as we go.