nord: auth: revert to signMessage for users
we've found that with certain wallets, the solana transaction framing hack does not work:

- solflare prepends two compute budget instructions to the message
  before signing, breaking our framing.
- solflare also has a bug where it says “wrong chain” when using devnet
  with our framing hack.
- glow straight up refuses to sign messages with our encoding, perhaps
  due to some validation around the blockhash.

there's also a strong precedent for solana apps not supporting sign in.
particularly salient examples are axiom.trade (uses turnkey internally)
and pump.fun (uses privy internally). we've, especially with the massive
testing matrix of various browser wallets, mobile wallets, hardware
wallets, and embedded wallets, among possibly others. signMessage is an
established primitive, with ledger just being an edge case.

it looks like that this is a concern being worked on. see:
https://github.com/anza-xyz/wallet-adapter/pull/1097

in other words, once Off-Chain Message Signing Format is formalized,
then the standard wallet adapter will offer it, where signing in with
ledger will be feasible. this will require an upgrade to change signing
formats, and will ultimately work and allow us to support ledger once
available. however we will not use hot wallets for admin signing,
because that would just cause more issues.

we've thus decided on the following:

- user sign-in will go back to using signMessage as before
- admin sign-in will stay with the signTransaction method

if ever the admin signing breaks with phantom, we always have the option
of bypassing and using the ledger sdk to sign directly. i've done this
before when testing and it is basically trivial to do.